requirements.txt
python-telegram-bot[job-queue]==22.1
anthropic>=0.40.0
elevenlabs>=1.0.0
pydub>=0.25.1
python-dotenv>=1.0.0
asyncpg>=0.30.0
